+++ anki/utils.py
@@ -18,6 +18,7 @@ from hashlib import sha1
 import platform
 import traceback
 import json
+from contextlib import contextmanager
 
 from anki.lang import _, ngettext
 
@@ -314,6 +315,13 @@ def namedtmp(name, rm=True):
 # Cmd invocation
 ##############################################################################
 
+@contextmanager
+def noBundledLibs():
+    oldlpath = os.environ.pop("LD_LIBRARY_PATH", None)
+    yield
+    if oldlpath is not None:
+        os.environ["LD_LIBRARY_PATH"] = oldlpath
+
 def call(argv, wait=True, **kwargs):
     "Execute a command. If WAIT, return exit code."
     # ensure we don't open a separate window for forking process on windows
@@ -327,7 +335,8 @@ def call(argv, wait=True, **kwargs):
         si = None
     # run
     try:
-        o = subprocess.Popen(argv, startupinfo=si, **kwargs)
+        with noBundledLibs():
+            o = subprocess.Popen(argv, startupinfo=si, **kwargs)
     except OSError:
         # command not found
         return -1
